refactor: view와 api 컨트롤러 패키지 분리
